requirements.txt
rpi-ws281x==4.2.4
pyserial==3.4
pyPS4Controller==1.2.5
joblib
rplidar-roboticia
mpu6050-raspberrypi==1.2
smbus
vl53l1x
scikit-learn
ArducamDepthCamera==0.1.3
numpy==1.26.4
lgpio==0.2.2.0
rpi-lgpio==0.6